check2

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -3,7 +3,7 @@ package lang.print.gaps.task5;
 public class NumberDivider {
     void divide(int toBeDivided) {
 
-        int y = toBeDivided/5;
+        double y = (double) toBeDivided/5;
         System.out.println(y);
 
     }
